IGNITE-12912 rewrite rules. Use separate classes
