+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/prepare/PlannerPhase.java
@@ -34,7 +34,8 @@ import org.apache.ignite.internal.processors.query.calcite.rule.TableModifyConve
 import org.apache.ignite.internal.processors.query.calcite.rule.UnionConverterRule;
 import org.apache.ignite.internal.processors.query.calcite.rule.UnionTraitsTraitsPropagationRule;
 import org.apache.ignite.internal.processors.query.calcite.rule.ValuesConverterRule;
-import org.apache.ignite.internal.processors.query.calcite.rule.logical.LogicalFilterRules;
+import org.apache.ignite.internal.processors.query.calcite.rule.logical.LogicalFilterMergeRule;
+import org.apache.ignite.internal.processors.query.calcite.rule.logical.LogicalFilterProjectTransposeRule;
 
 import static org.apache.ignite.internal.processors.query.calcite.prepare.IgnitePrograms.cbo;
 import static org.apache.ignite.internal.processors.query.calcite.prepare.IgnitePrograms.decorrelate;
@@ -74,8 +75,8 @@ public enum PlannerPhase {
                 ProjectConverterRule.INSTANCE,
                 ProjectTraitsPropagationRule.INSTANCE,
                 FilterConverterRule.INSTANCE,
-                LogicalFilterRules.FILTER_MERGE_RULE,
-                LogicalFilterRules.FILTER_PROJECT_TRANSPOSE_RULE,
+                LogicalFilterMergeRule.INSTANCE,
+                LogicalFilterProjectTransposeRule.INSTANCE,
                 FilterTraitsPropagationRule.INSTANCE,
                 TableModifyConverterRule.INSTANCE,
                 UnionMergeRule.INSTANCE,
